[Core] GetVersion returns ushort now instead uint [Views] GroupChangedTracker set up version as MaxValue by default

+++ Runtime/Core/Ents/Ents.Ent.RO.cs
@@ -21,7 +21,7 @@ namespace ME.BECS {
         public readonly ref readonly World World => ref this.ent.World;
 
         [INLINE(256)]
-        public readonly uint GetVersion(uint groupId) => this.ent.GetVersion(groupId);
+        public readonly ushort GetVersion(uint groupId) => this.ent.GetVersion(groupId);
 
         [INLINE(256)]
         public readonly ulong ToULong() => this.ent.ToULong();
